pyproject.toml
[project]
name = "omniagent"
version = "0.1.0"
description = "OmniAgent - Your self-evolving AI agent framework"
readme = "README.md"
license = {text = "GPL-3.0-only"}
requires-python = ">=3.11"
authors = [
    {name = "OmniAgent Contributors"},
]
keywords = ["ai-agent", "llm", "function-calling", "self-evolving", "reflexion", "multi-agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "aiohttp>=3.8.0",
    "click>=8.1.7",
    "pydantic>=2.5.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.1",
    "websockets>=12.0",
    "psutil>=5.9.0",
    "rich>=13.7.0",
    "openai>=1.60.0",
    "anthropic>=0.40.0",
    "beautifulsoup4>=4.12.0",
    "prompt-toolkit>=3.0.0",
]

[project.optional-dependencies]
feishu = ["lark-oapi>=1.5.0"]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "mypy>=1.7.1",
    "black>=23.12.0",
    "isort>=5.13.2",
]

[project.scripts]
omniagent = "omniagent.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["omniagent*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
